(Backend) Added filtering by query & Fixed ReponseDtos and Mappers to return better info for Article

+++ src/main/java/irinakjoseva/vecnamoda/dto/response/CategoryResponseDto.java
@@ -6,6 +6,6 @@ public class CategoryResponseDto {
 
     public String name;
 
-    public Long parentCategoryId;
+    public Integer parentCategoryId;
 
 }
